# ollama-distributed/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OllamaMax Distributed API
  description: |
    OllamaMax Distributed is an enterprise-grade distributed AI platform that transforms 
    the single-node Ollama architecture into a horizontally scalable, fault-tolerant system.
    
    ## Authentication
    
    The API uses JWT-based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <JWT_TOKEN>
    ```
    
    ## Base URLs
    
    - **API Server**: `http://localhost:8080`
    - **Web Interface**: `http://localhost:8081`
    - **Metrics**: `http://localhost:9090`
    
    ## Rate Limiting
    
    Default rate limit: 1000 requests per minute per client.
    
  version: 1.0.0
  contact:
    name: OllamaMax Support
    url: https://github.com/khryptorgraphics/ollamamax
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.ollamamax.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      summary: User authentication
      description: Authenticate user and receive JWT token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "admin"
              password: "secure-password"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/models:
    get:
      summary: List available models
      description: Get a list of all available AI models in the cluster
      tags:
        - Models
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of models per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
  /api/v1/models/{name}:
    get:
      summary: Get model details
      description: Get detailed information about a specific model
      tags:
        - Models
      parameters:
        - name: name
          in: path
          required: true
          description: Model name
          schema:
            type: string
          example: "llama2:7b"
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDetails'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete model
      description: Remove a model from the cluster
      tags:
        - Models
      parameters:
        - name: name
          in: path
          required: true
          description: Model name
          schema:
            type: string
      responses:
        '200':
          description: Model deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/UserInfo'
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
    ModelsResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelSummary'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ModelSummary:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
          format: int64
        digest:
          type: string
        modified_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [available, downloading, error]
    ModelDetails:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
          format: int64
        digest:
          type: string
        format:
          type: string
        family:
          type: string
        parameter_size:
          type: string
        quantization_level:
          type: string
        capabilities:
          type: array
          items:
            type: string
        modified_at:
          type: string
          format: date-time
        download_count:
          type: integer
        nodes:
          type: array
          items:
            type: string
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
        details:
          type: string
        timestamp:
          type: string
          format: date-time
...